requirements.txt
boto3
PyMySQL
cryptography
opencv-python-headless
cv2_tools
requests